lib: at_host: Send also potential termination characters
Include also characters like <CR> and <LF> in the buffer that's
sent to the modem. Certain AT commands, like AT+CMGS, use <CR>
as part of the command format.

This patch also has the side-effect of letting the AT host
detect closing quote character also in cases of buffer overflow.
Note that the closing quote will not actually be included in
the buffer, only detected and allow for the command to be sent
if the configured termination character(s) is/are detected.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>