net: lib: nrf_cloud: Remove transport dependency for A-GPS and P-GPS
Currently, nRF Cloud library must be enabled in its entirety to be
able to use A-GPS and P-GPS parsing functions.
This can be problematic when a user wants to have its own cloud
solution but still receive data from nRF Cloud location services
over its own transport and process it.

This patch removes the dependency on nRF Cloud library for
A-GPS and P-GPS processing functions.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>