drivers: wifi: Remove workaround for RPU power on
We have now found the root cause of the issue as to why 10ms delay is
needed after asserting buck enable, all the nrf7002 shields have a pin
conflict with (CPUNET UART0 TX, RX) and (buck_en, iovdd), they use D1
and D0 in the Arduino header.

So, during boot the banner (and another print) in CPUNET UART is
conflicting with buck_end and the GPIO goes wobbly, the prints take
about 8ms, hence the need for 10ms sleep.

Now that we have disabled CPUNET UART0 for all nrf7002 EKs, this
workaround can be removed.

Fixes SHEL-685.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>